refactor(home): update home to pull hero data
