+++ next.config.js
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  images: {
+    domains: ['cdn.jsdelivr.net'],
+  },
 }
 
 module.exports = nextConfig
