+++ pages/_app.tsx
@@ -1,9 +1,12 @@
 import * as React from "react";
 import Head from "next/head";
 import { AppProps } from "next/app";
+
 import { ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
 import { CacheProvider, EmotionCache } from "@emotion/react";
+import { SWRConfig } from "swr";
+
 import theme from "../utils/theme";
 import emotionCache from "../utils/emotion-cache";
 
@@ -21,9 +24,14 @@ export default function MyApp(props: MyAppProps) {
         <meta name="viewport" content="initial-scale=1, width=device-width" />
       </Head>
       <ThemeProvider theme={theme}>
-        {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
         <CssBaseline />
-        <Component {...pageProps} />
+        <SWRConfig
+          value={{
+            fetcher: (resource, init) => fetch(resource, init).then((res) => res.json()),
+          }}
+        >
+          <Component {...pageProps} />
+        </SWRConfig>
       </ThemeProvider>
     </CacheProvider>
   );
