+++ pages/index.tsx
@@ -1,73 +1,17 @@
 import type { NextPage } from "next";
-import Head from "next/head";
-import Image from "next/image";
-import Link from "next/link";
-import styles from "../styles/Home.module.css";
+import useSWR from "swr";
+
+import List from "../components/List";
 
 const Home: NextPage = () => {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to{" "}
-          <Link href="https://nextjs.org">
-            <a>Next.js!</a>
-          </Link>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <Link href="https://nextjs.org/docs">
-            <a className={styles.card}>
-              <h2>Documentation &rarr;</h2>
-              <p>Find in-depth information about Next.js features and API.</p>
-            </a>
-          </Link>
-
-          <Link href="https://nextjs.org/learn">
-            <a className={styles.card}>
-              <h2>Learn &rarr;</h2>
-              <p>Learn about Next.js in an interactive course with quizzes!</p>
-            </a>
-          </Link>
-
-          <Link href="https://github.com/vercel/next.js/tree/canary/examples">
-            <a className={styles.card}>
-              <h2>Examples &rarr;</h2>
-              <p>Discover and deploy boilerplate example Next.js projects.</p>
-            </a>
-          </Link>
-
-          <Link href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app">
-            <a className={styles.card}>
-              <h2>Deploy &rarr;</h2>
-              <p>Instantly deploy your Next.js site to a public URL with Vercel.</p>
-            </a>
-          </Link>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <Link href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app">
-          <a target="_blank" rel="noopener noreferrer">
-            Powered by{" "}
-            <span className={styles.logo}>
-              <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-            </span>
-          </a>
-        </Link>
-      </footer>
-    </div>
+  const { data, error } = useSWR(
+    "https://cdn.jsdelivr.net/gh/akabab/superhero-api@0.3.0/api/all.json"
   );
+
+  if (error) return <>An error has occurred.</>;
+  if (!data) return <>Loading...</>;
+
+  return <List items={data} />;
 };
 
 export default Home;
